refactor(react-library): delete library.js — the contract moves to library-globals.js
7,446 lines gone. a reachability pass over all 154 remaining functions found
74 already dead and the rest reachable only from inside the file — except the
five names other classic scripts still call by global name, which moved
VERBATIM into static/library-globals.js: artistDetailPageState + the
back-label stack, navigateToArtistDetail, playLibraryTrack,
_updateSidebarLibraryBreadcrumb, _handoffLibrarySearchToEnhancedSearch.

one deliberate change inside navigateToArtistDetail: it no longer cancels an
inline edit or removes #enhanced-manual-match-overlay. both are React-owned
now, and removing a React-rendered node out from under its owner corrupts the
tree on unmount — the route change unmounts it properly.

playTrackByMetadata did NOT move: React ported it into
-artist-detail.top-tracks.ts and the vanilla copy had no callers left.

the guards updated deliberately, as test_artist_detail_cross_file_contract's
own header demanded: it pins library-globals.js now, the react-handoff guard
follows the Watch All seam into React, and the release-classifier parity test
freezes the vanilla regexes as literals since its vanilla half is gone. the
duplicate-id allowlist is empty and the last two static twins are deleted.
